fix: Input/Output Polymorphism
refactor with a base class of Port for common code.
